ump cyclers at build time
